Avoid use of ansible_host and machine UUID
Using ansible_host isn't very portable, such as when you're deploying
into a cloud environment. Instead of using a value that was assigned
to ansible_host in a locally generated inventory file, attempt to use
the default IPv4 address of the host.

In the future we'll probably want to make this configurable where  you
could use the IP address of whatever interface you specify (with a
default of this value if no interfaces are specified).

Clean up topology.json template for gk-deploy to use the hostname
instead of the machine UUID. When the kube join command is run it
likely uses the hostname, so we should template with that value.

Related to #55