Drop the use of all_vms
When using this role outside of a local (or remote) virtual environment,
the use of all_vms group can cause some issues and confusion (primarily
when you're trying to deploy in something like a cloud environment fronted
by AWX).

This change replaces the use of all_vms with a call to both master and
nodes groups, and updates templates that generate local configuration to
now duplicate the master and nodes vars sections. Documentation has also
been updated to reflect this change.

Closes #80